backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
langchain==1.2.10
langchain-community==0.4.1
langchain-core==1.2.14
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.1
faiss-cpu==1.13.2
huggingface_hub==0.36.2
sentence-transformers==5.2.3
python-dotenv==1.2.1
pypdf==6.7.2
# Auth & database
motor>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
python-jose[cryptography]>=3.3,<4.0
